list status when create lister
